requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
drf-yasg==1.21.7
environ==1.0
Jinja2==3.1.3
pillow==10.3.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
